Move /structured under /experimental/structured and move out dataset config from ModelPT dataset config
