+++ examples/asr/experimental/structured/speech_to_text_hybrid.py
