+++ examples/asr/experimental/structured/speech_to_text_structured.py
